# docs/openapi/api.yaml
openapi: 3.0.3
info:
  title: Remit-Scout API
  version: 1.0.0
  description: |
    Remit-Scout API for comparing money transfer providers and accessing remittance data.
    
    ## Authentication
    - Public endpoints: Rate-limited by IP
    - Protected endpoints: Require JWT Bearer token unless the route explicitly supports `X-API-Key`
    - `X-API-Key` is route-scoped and authoritative when present
      - Retail user API keys are created under `/api/v1/me/api-keys` and can access only explicit retail API-key routes, subject to scope checks
      - Institutional API keys are admin-managed, corridor-aware, and limited to explicit B2B routes
      - Presenting `X-API-Key` on an unsupported route returns `403 api_key_route_not_allowed`
      - Invalid keys return `401 invalid_api_key`; revoked retail keys return `401 revoked_api_key`
    
    ## Rate Limiting
    - Public: 100 requests per window (default 1 minute)
    - Authenticated: 500 requests per window
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.remit-scout.com
    description: Production
paths:
...
  /api/v1/compliance/status:
    get:
      summary: Get institutional compliance and launch-readiness status
      description: |
        Returns public-facing compliance metadata and the institutional data-maturity gate.
        This surface is used by the institutional waitlist pages and does not require authentication.
      tags:
        - Institutions
      responses:
        '200':
          description: Compliance and launch-readiness status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceStatusResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/me:
    get:
      summary: Get current user information
      description: |
        Returns raw purchased plan state, effective capability state, billing summary,
        entitlements, and usage. `plan` is the stored purchased tier. `plan_effective`
        is the canonical capability plan after lifecycle fallback and admin overrides.
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/me/api-keys:
    get:
      summary: List retail enterprise user API keys
      description: |
        Returns retail user API keys for the current authenticated enterprise user.
        Institutional API keys are provisioned through admin institutional client flows and are not listed here.
        Inactive paid plans fail with canonical `plan_inactive` detail fields.
      tags:
        - API Keys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API keys list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create retail enterprise user API key
      description: |
        Creates a retail user API key for the current authenticated enterprise user.
        Supported scopes are `indices:read`, `corridors:read`, and `exports:read`.
        Unknown scopes and legacy values such as `tier:1` are rejected.
        Inactive paid plans fail with canonical `plan_inactive` detail fields.
      tags:
        - API Keys
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                scopes:
                  description: Optional retail API-key scopes. Defaults to `indices:read` and `corridors:read`.
                  type: array
                  items:
                    type: string
                    enum: [indices:read, corridors:read, exports:read]
      responses:
        '200':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/billing/checkout-session:
    post:
      summary: Create Stripe checkout session
      description: Creates a Stripe checkout session for upgrading to Plus plan
      tags:
        - Billing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_code:
                  type: string
                  enum: [plus, enterprise]
                billing_interval:
                  type: string
                  enum: [month, year]
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                  session_id:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already has an active or trialing Plus subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/billing/webhook:
    post:
      summary: Stripe webhook handler
      description: Handles Stripe webhook events (internal use only)
      tags:
        - Billing
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Route-scoped API key for supported retail user and institutional consumers. Unsupported routes fail with `403 api_key_route_not_allowed`.
  schemas:
...
    ComplianceStatusResponse:
      type: object
      required:
        - updated_at
        - certifications
        - institutional_data_maturity
        - privacy_controls
      properties:
        updated_at:
          type: string
          format: date-time
        certifications:
          type: object
          properties:
            gdpr:
              type: string
            ccpa:
              type: string
            soc2_type_ii:
              type: object
              properties:
                status:
                  type: string
                report_state:
                  type: string
                report_date:
                  type: string
                  nullable: true
                report_url:
                  type: string
                  nullable: true
                expires_on:
                  type: string
                  nullable: true
        institutional_data_maturity:
          $ref: '#/components/schemas/InstitutionalDataMaturity'
        privacy_controls:
          type: object
          properties:
            data_minimization:
              type: string
            ip_handling:
              type: string
            session_identity:
              type: string
...
    MeResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        user:
          type: object
          properties:
            user_id:
              type: string
            email:
              type: string
            name:
              type: string
              nullable: true
            role:
              type: string
              nullable: true
            app_role:
              type: string
              nullable: true
            is_admin:
              type: boolean
            mfa_verified:
              type: boolean
        plan:
          $ref: '#/components/schemas/MePlanResponse'
        plan_effective:
          $ref: '#/components/schemas/MePlanEffectiveResponse'
        billing:
          $ref: '#/components/schemas/MeBillingResponse'
        entitlements:
          $ref: '#/components/schemas/MeEntitlementsResponse'
        usage:
          type: object
          additionalProperties: true
    PlanCode:
      type: string
      enum: [free, plus, enterprise]
...
    PlanLifecycleState:
      type: string
      enum: [active, trialing, scheduled_cancel, past_due, canceled, unpaid, incomplete_expired, expired, inactive]
    PlanRecoveryAction:
      type: string
      enum: [none, billing_portal, upgrade]
    MePlanResponse:
      type: object
      properties:
        plan_code:
          $ref: '#/components/schemas/PlanCode'
        status:
          $ref: '#/components/schemas/PlanStatus'
...
    ApiKeyListResponse:
      type: object
      properties:
        success:
          type: boolean
        keys:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyRecord'
    ApiKeyCreateResponse:
      type: object
      properties:
        success:
          type: boolean
        api_key:
          $ref: '#/components/schemas/ApiKeyRecord'
        token:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: object
              properties:
                plan_failure:
                  type: string
                required_plan:
                  $ref: '#/components/schemas/PlanCode'
                capability:
                  type: string
                lifecycle_state:
                  $ref: '#/components/schemas/PlanLifecycleState'
                recovery_action:
                  $ref: '#/components/schemas/PlanRecoveryAction'
                purchased_plan:
                  $ref: '#/components/schemas/PlanCode'
                effective_plan:
                  $ref: '#/components/schemas/PlanCode'
              additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication. API-key callers may receive `invalid_api_key` or `revoked_api_key`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - Insufficient permissions. API-key callers may receive `api_key_route_not_allowed`, `insufficient_scope`, `corridor_not_allowed`, `institutional_inactive`, or `institutional_only`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...